All estimate methods now take in users and items id, in that order.
Also put some stuff that were more relevent in the Trainset class.

It's up to the algo to choose if it wants to be symmetric or not. I created a
base class, which I don't like. Plus the compute_similarities() function is now
dependent on this class, even though it's still in AlgoBase.
Will need to back to it, but my head is gonna explode right now.