Slightly changed setup.py machinery.
Ok, things are a bit clearer now. We do something similar to scikit-learn: when
installinf recsys from PyPI, cython is not required so the .so need to be
shipped in the wheel. When using python setup.py install, both numpy and cython
need to be installed. We could have an option for cython (and distrbute .c
files when cythonization is not available), but we have the numpy bootstrap
problem anyway: it's needed as a dependency BUT we also need it for cythonizing
stuff at the installation stage. There are some workaround but I don't like
them. So fuck it, if you want to use setup.py, you'll need to have both numpy
and cython installed.